Create torch.func, add APIs from torch._functorch.eager_transforms
This PR creates the torch.func APIs. (If we don't like the name, we have
until the next release to change them).

The APIs are a thin wrapper around the implementation of the APIs in
torch._functorch.eager_transforms + documentation. This is to satisfy
PyTorch public API rules.

The documentation was copy-pasted from their original locations, with
slight modifications:
- I renamed all instances of functorch to torch.func.

Test Plan:
- wait for tests

ghstack-source-id: a068eeb092d0dd7d8c09cfc8aa38621f31cd1cea
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90331